Added LNC-Linker Coin Token to Defaults
Linker Coin is a Cryptocurrency based on Ethereum Technology ERC20 protocol.
This is ourCoin's address:
0xf7d58e6720696B8f46Dc9270443D8beaeb8d9775
This is our website: www.linkercoin.com/en
As of now, we have already surpassed the expectations for our third pre-sale (which took place on 11/18) and we are now dealing with our Crowd Sale which started on November 25 2017
We are establishing a Network of Networks Network of Networks, a Cryptocurrency Exchange where any cryptocurrency could potentially trade and gain liquidity. Through this exchange, liquidity could be provided to all participating cryptocurrencies. Our cryptocurrency is based on Ethereum Technology and through the market-making expertise of some of its cofounders will create an environment of an Efficient Cryptocurrency Market. In this Cryptocurrency Exchange, cryptocurrencies could be exchanged with other cryptocurrencies as well as fiat currencies. To further strengthen the liquidity of the participating entities and in order to have a complete market, a derivatives (plain vanilla instruments such as options, futures, and swaps with certain cryptocurrencies as underlying asset) market is also planned to be launched soon.
We are a team of Expert Market Makers, Traders and Designers of Trading Strategies.