Update tokens-eth.json
Rainbow Currency Foundation is a Non-Profit Corporation, registered in Nevada
Business ID NV20171765116.

You simply hold a smart new cryptocurrency, that has been built to make global and local payment processes safer, cheaper, and easier. Purchases happen internally for this ... for example, it makes no sense for a customer wanting to buy a coffee to have to create a transaction on the blockchain for it ... this way payment is instant and has zero fees.

Thanks to a unique 2-wallet solution, the whole currency is always fully transparent, decentralized, and safe in the blockchain, while everyday payments for shopping and services can be made online and at POS fast, secure, and without fees.

To be qualified for online and local payments, the currency holder must go through a verification process to identify themselves, while only accredited business partners are allowed to receive these payments. Thus, we make it difficult and uncomfortable to participate for the bad guys.

The token is on a Standard ERC20 smart contract avoiding misuse. Contract Address: 0xfbd0d1c77b501796a35d86cf91d65d9778eee695

Rainbow Currency is NOT an investment into a certain project. When you own Rainbow Currency, you are not holding any partnership, stock, or revenue share rights.

There is never any expectation of profits, as it is not sold as an investment. Pure demand will generate an increase in token pricing and this will happen organically. Yes of course, we will have advertising campaigns and they will be managed by the company, as with any business, exposure is a good thing but it is used to promote products and services and never about the token.