add ENTRP to tokens
The Hut34 Project is building a blockchain based protocol to facilitate the open exchange and monetization of data, information and digital services. It’s where A.I does business: An ecosystem of chatbots, A.I. and IoT services where developers can easily connect and route queries, receive resolution of these queries, add utility to their end users, and be fairly rewarded for this. The Entropy Token is a utility token, that powers the Hut34 Protocol and Network.