Added CLL Token to Defaults
Token is currently trading here: https://forkdelta.github.io/#!/trade/CLL-ETH

Description:
CryptoLiveLeak is a leading cryptocurrency news and media platform. We have integrated the CLL token with our platform to innovative ways the audience and media interact. The CLL Token has multiple benefits including exclusive discounts when participating in future ICOs from founder or through the CryptoLiveLeak platform. We have a large dedication to giving the CLL tokens to those who interact with our platform as show by our Airdrop and Giveaways. Token holders can also accumulate a LEAK SCORE to gain additional incentives when interacting with the CryptoLiveLeak platform. We aim to innovate and disrupt in a variety of industries and want our early supporters to gain with us by holding CLL token. This all occurs within a trustless environment where all participants have transparency. The platform will evolve as we progress with IBO efforts. The CryptoLiveLeak team will also look to expand and grow as the network matures. Large budget for development and marketing. The CryptoLiveLeak (CLL) ERC20 token will give holders an unique way of participating in the evolution of our vision of blockchain tech including transforming the payment and exchanges industries.