Added PTC
ParrotCoin is a coin made with love for Parrots. With ParrotCoins we keep our Parrot Park up and running. Here we keep our most valued Parrots which are on the edge of extinction. We try to let them breed and release their youngs in the wild again when they are grown up. This way we help the nature and keep those amazing parrots alive. With this coin you can watch our LiveCams and keep up to date with the life of our Parrots. The value of the ParrotCoin will rise up by time as there is a maximum cap available.

For more information please look at our website or contact us directly at: info@parrotcoin.club