Added HUR Token to Defaults
HUR tokens are ERC20 compliant utility tokens that can be used only on the Hurify platform. The HUR token is the common medium of exchange for smart contract transactions enabled on the Hurify platform.
Hurify is a platform to facilitate the development of the Internet of Things (“IoT”) projects, including the matching of project owners with project managers and project development subcontractors, and hardware and software vendors within the IoT ecosystem. We combine global IoT development talent matching, smart contract-based IoT project management and e-commerce to facilitate IoT project development.