Added MHLK Token to Defaults
I just wanted to make it sure that we are both on the same page. The project is not an ICO. The project is to distribute the wealth of the Maharlika Trust. We are utilizing both ERC721 and ERC20 standard technology to accomplish the goals of our project. We used (ERC721) to digitize the assets under each Maharlika Asset Coin (MHKC) which will be used to back our (ERC20) Maharlika Coin (MHLK).