0xe4e2daf5f7f0c1c35df922c5d9340913edecc8e8.json
Mnoer or MNR is more than just a cryptocurrency. It was found to solve the problem of alternative energy. Mnoer's mission is to take advantage of transparency and consistency in blockchain technology and to start a solar power company to provide the energy requirements for people around the world. Our idea is now on Blockchain and ready for investors!
Our MNR currency project will be ready to save energy and get paid in return. Profits will be shared by the owners. That is, each portion of the profits will be shared among MNR holders monthly.